+++ packages/interface/src/peer-store/index.ts
@@ -166,7 +166,7 @@ export interface PeerStore {
    * })
    * ```
    */
-  forEach: (fn: (peer: Peer) => void, query?: PeerQuery) => Promise<void>
+  forEach(fn: (peer: Peer) => void, query?: PeerQuery): Promise<void>
 
   /**
    * Returns all peers in the peer store.
@@ -179,7 +179,7 @@ export interface PeerStore {
    * }
    * ```
    */
-  all: (query?: PeerQuery) => Promise<Peer[]>
+  all(query?: PeerQuery): Promise<Peer[]>
 
   /**
    * Delete all data stored for the passed peer
@@ -197,7 +197,7 @@ export interface PeerStore {
    * // []
    * ```
    */
-  delete: (peerId: PeerId) => Promise<void>
+  delete(peerId: PeerId): Promise<void>
 
   /**
    * Returns true if the passed PeerId is in the peer store
@@ -212,7 +212,7 @@ export interface PeerStore {
    * // true
    * ```
    */
-  has: (peerId: PeerId) => Promise<boolean>
+  has(peerId: PeerId): Promise<boolean>
 
   /**
    * Returns all data stored for the passed PeerId
@@ -224,7 +224,7 @@ export interface PeerStore {
    * // { .. }
    * ```
    */
-  get: (peerId: PeerId) => Promise<Peer>
+  get(peerId: PeerId): Promise<Peer>
 
   /**
    * Adds a peer to the peer store, overwriting any existing data
@@ -237,7 +237,7 @@ export interface PeerStore {
    * })
    * ```
    */
-  save: (id: PeerId, data: PeerData) => Promise<Peer>
+  save(id: PeerId, data: PeerData): Promise<Peer>
 
   /**
    * Adds a peer to the peer store, overwriting only the passed fields
@@ -250,7 +250,7 @@ export interface PeerStore {
    * })
    * ```
    */
-  patch: (id: PeerId, data: PeerData) => Promise<Peer>
+  patch(id: PeerId, data: PeerData): Promise<Peer>
 
   /**
    * Adds a peer to the peer store, deeply merging any existing data.
@@ -263,7 +263,7 @@ export interface PeerStore {
    * })
    * ```
    */
-  merge: (id: PeerId, data: PeerData) => Promise<Peer>
+  merge(id: PeerId, data: PeerData): Promise<Peer>
 
   /**
    * Unmarshal and verify a signed peer record, extract the multiaddrs and
@@ -278,5 +278,5 @@ export interface PeerStore {
    * await peerStore.consumePeerRecord(buf, expectedPeer)
    * ```
    */
-  consumePeerRecord: (buf: Uint8Array, expectedPeer?: PeerId) => Promise<boolean>
+  consumePeerRecord(buf: Uint8Array, expectedPeer?: PeerId): Promise<boolean>
 }